Revise the way Xcode Project Integration works, in order to let cocoapods do more of the heavy lifting.
Basically, just add the Pods project to the Unity-iPhone project as a "subproject" (the way you used to do multiproject builds before workspaces), and add it to the build of the main target.  Also, don't suppress the pod command's normal project integration (i.e., don't add the line "install! 'cocoapods', :integrate_targets => false" to the Podfile) since that step does all of the work in setting up the build dependencies between the Unity-iPhone project and the Pods project.  That includes propagating build settings (search paths, flags, etc) from the pods into the main project.  (I believe this means the "bitcodeEnabled" attribute is now superfluous, but I didn't disturb the implementation of it for backwards compatibility.)

This change, aside from simplifying the code quite a bit, also lifts some restrictions that the old implementation had.  Local (i.e., "Development") pods that are located outside the Pods subdirectory are now supported.  Also, the resources in source pods are properly included in the app build.

Also leveraged some static calls provided by Unity's PBXProject class for finding the path project and main target name, so that logic is no longer needed in this code.